% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{datexemplo}
\alias{datexemplo}
\title{Dados de temperatura da superficie dos oceanos}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 192240 rows and 4 columns.
}
\source{
https://www.ncei.noaa.gov/pub/data/cmb/ersst/v5/netcdf/
}
\usage{
datexemplo
}
\description{
data.table contendo informacoes de temperatura dos oceanos exemplo no periodo 2020-01 a 2020-12
}
\details{
\describe{
\item{DATE}{data do registro}
\item{LON}{longitude do registro}
\item{LAT}{latitude do registro}
\item{SST}{temperatura na superficie do oceano}
}
}
\keyword{datasets}
