% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{leitura}
\alias{leitura}
\alias{leFROMdir}
\alias{leFROMurl}
\title{Leitura dos dados de sst}
\usage{
leFROMdir(dir)

leFROMurl(janela, url, outdir)
}
\arguments{
\item{dir}{diretorio onde se encontram os arquivos .nc no formato ersst.v5.AAAAMM.nc, onde
\emph{AAAA} indica o ano e \emph{MM} o mes referente as informacoes}

\item{janela}{vetor de dois elementos no formato "AAAAMM" indicando a janela de tempo dos dados
que devem ser baixados}

\item{url}{url de onde baixar os dados. Se nao forncida, sera usado
\url{https://www.ncei.noaa.gov/pub/data/cmb/ersst/v5/netcdf/}}

\item{outdir}{opcional; diretorio onde salvar os arquivos baixados. Caso nao seja fornecido, os
arquivos serao salvos em um diretorio temporario na raiz padrao do sistema operacional}
}
\value{
data.table contendo informacoes formatadas para uso
\describe{
\item{DATE}{data do registro}
\item{LON}{longitude do registro}
\item{LAT}{latitude do registro}
\item{SST}{temperatura na superficie do oceano}
}
}
\description{
Funcoes para leitura de dados netCDF ja baixados ou a partir da internet
}
\examples{

# LEITURA DE UM DIRETORIO LOCAL CONTENDO ARQUIVOS ---------------------

# usando o diretorio dos dados inclusos no pacote
diretorio <- system.file("extdata", package = "oceantemps")

# o pacote so inclui dois meses (2020-01 e 2020-02)
list.files(diretorio)

# realiza a leitura
leFROMdir(diretorio)

# LEITURA DA MESMA JANELA DE DADOS A PARTIR DA URL PADRAO -------------

leFROMurl(janela = c("202001", "202002"))

}
\seealso{
\code{\link{plotamapa}} para metodos de visualizacao deste dado
}
