% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plotamapa}
\alias{plotamapa}
\title{Plot da temperatura na superficie}
\usage{
plotamapa(dado, data, outfile)
}
\arguments{
\item{dado}{o dado do qual extrair serie. Deve ter formato igual ao retornado por
\code{\link{leFROMdir}} ou \code{\link{leFROMurl}}}

\item{data}{escalar string no formato AAAAMM indicando o ano-mes que deve ser plotado}

\item{outfile}{opcional, indicando um caminho de arquivo para escrita. Ver Exemplos}
}
\value{
objeto ggplot contendo grafico criado ou vazio no caso de \code{outfile} fornecido
}
\description{
Funcao facilitadora para visualizacao do dado de temperatura na superficie
}
\examples{

# plot do ano-mes 202001
plotamapa(datexemplo, "202001")

\dontrun{
# escrita do arquivo em um diretorio qualquer
plotamapa(datexemplo, "202001", "C:/Users/usuario_qualquer/Imagens/plot.png")
}

}
